+++ src/app.module.ts
@@ -3,6 +3,7 @@ import { ConfigModule } from '@nestjs/config';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { validateEnv } from './config/env.validation';
+import { BrandsModule } from './brands/brands.module';
 import { DatabaseModule } from './database/database.module';
 import { HealthModule } from './health/health.module';
 import { ProductsModule } from './products/products.module';
@@ -19,6 +20,7 @@ import { ProductsModule } from './products/products.module';
     }),
     DatabaseModule,
     HealthModule,
+    BrandsModule,
     ProductsModule,
   ],
   controllers: [AppController],